+++ client/src/containers/NavBarContainer.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import NavBar from '../components/NavBar';
+
+let NavBarContainer = React.createClass({
+  render() {
+    return (
+      <NavBar/>
+    );
+  }
+});
+
+export default NavBarContainer;